packages/whistleblower/requirements.txt
pydbus
python-telegram-bot
pygobject
